pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "renderbrains-docs"
version = "0.1.0"
description = "RenderBrains documentation toolchain"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pyyaml>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["scripts"]
